requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-admin-tools==0.9.0
django-bootstrap4==1.1.1
django-bower==5.2.0
django-controlcenter==0.3.0
django-heroku==0.3.1
django-jsonfield==1.4.0
django-nvd3==0.9.7
django-pkgconf==0.4.0
django-qsstats-magic==1.1.0
djcacheutils==3.0.0
gunicorn==20.0.4
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
python-dateutil==2.8.1
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
Unidecode==1.1.1
whitenoise==5.0.1
